requirements.txt
python-telegram-bot==21.6
openai==1.58.1
pydub==0.25.1
